Add functional ground class and modify player class to run on it
Creates ground class as a list of heights and built methods into it
which advance the ground, generate new terrain, and draw it. Changed
the player class so instead of landing on the bottom of the window it
will land on the ground. Game is now working to run along ground and
jump as you go.