Add rock class and rock group, build game over event, coin generation
Added coin generator in main loop that keeps track of time and adds
a new coin object at regular intervals (but not in a gap). Built game
over screen which runs when the main loop is exited and shows the score.
Built rock class and collision check with rocks that will end the game,
but nothing yet is generating rocks.